+++ NEWS.md
@@ -1,4 +1,13 @@
-# metropolis v0.1.5
+# metropolis 
+
+# v0.1.8
+## Major changes
+- Better error handling in likelihood functions
+
+## Bug fixes
+- Fixed bug in log-Jacobian calculation for Gaussian model
+
+# v0.1.5
 ## Major changes
 - First release on CRAN